{"schema":"decodex/commit/1","summary":"Modularize docs OKF checks","authority":"XY-1115"}
